Add spacing and style formatting to form content on mid-right container.
Prior to this change, the form elements in the mid-right panel had
default styles. This meant they were not spaced well and were
not visually appealing. This change spaces those inputs and puts
them into columns. It adds fonts to the labels and form heading,
changing font weight and font color.